backend/requirements.txt
python-dotenv
google-generativeai
langchain-community
langchain-core
chromadb
sentence-transformers
langchain-huggingface
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
pyjwt[crypto]
passlib[bcrypt]
python-multipart
email-validator
httpx
